checking in more progress
