+++ watertap/unit_models/reverse_osmosis_0D.py
@@ -30,6 +30,7 @@ from idaes.core.util.exceptions import ConfigurationError
 from idaes.core.solvers import get_solver
 from idaes.core.util.misc import add_object_reference
 import idaes.core.util.scaling as iscale
+from watertap.core import MembraneChannel0D
 from watertap.core.util.initialization import check_solve, check_dof
 from watertap.unit_models._reverse_osmosis_base import (
     ConcentrationPolarizationType,
@@ -57,22 +58,9 @@ class ReverseOsmosisData(_ReverseOsmosisBaseData):
 
     CONFIG = _ReverseOsmosisBaseData.CONFIG()
 
-    def build(self):
-        """
-        Build the RO model.
-        """
-        # Call UnitModel.build to setup dynamics
-        super().build()
-
-        # for quacking like 1D model -> 0. is "in", 1. is "out"
-        self.length_domain = Set(
-            ordered=True, initialize=(0.0, 1.0)
-        )  # inlet/outlet set
-        add_object_reference(self, "difference_elements", self.length_domain)
-        self.first_element = self.length_domain.first()
-
-        # Build control volume for feed side
-        self.feed_side = ControlVolume0DBlock(
+    def _add_membrane_channel(self):
+        # Build membrane channel control volume
+        self.membrane_channel = MembraneChannel0D(
             default={
                 "dynamic": False,
                 "has_holdup": False,
@@ -80,238 +68,6 @@ class ReverseOsmosisData(_ReverseOsmosisBaseData):
                 "property_package_args": self.config.property_package_args,
             }
         )
-
-        self.feed_side.add_state_blocks(has_phase_equilibrium=False)
-
-        self.feed_side.add_material_balances(
-            balance_type=self.config.material_balance_type, has_mass_transfer=True
-        )
-
-        self.feed_side.add_energy_balances(
-            balance_type=self.config.energy_balance_type, has_enthalpy_transfer=True
-        )
-
-        self.feed_side.add_momentum_balances(
-            balance_type=self.config.momentum_balance_type,
-            has_pressure_change=self.config.has_pressure_change,
-        )
-
-        # for quacking like 1D model
-        add_object_reference(
-            self.feed_side,
-            "properties",
-            {
-                **{
-                    (t, 0.0): self.feed_side.properties_in[t]
-                    for t in self.flowsheet().config.time
-                },
-                **{
-                    (t, 1.0): self.feed_side.properties_out[t]
-                    for t in self.flowsheet().config.time
-                },
-            },
-        )
-
-        # Add additional state blocks
-        tmp_dict = dict(**self.config.property_package_args)
-        tmp_dict["has_phase_equilibrium"] = False
-        tmp_dict["parameters"] = self.config.property_package
-        tmp_dict["defined_state"] = False  # these blocks are not inlets
-        # Build permeate side
-        self.permeate_side = self.config.property_package.state_block_class(
-            self.flowsheet().config.time,
-            self.length_domain,
-            doc="Material properties of permeate along permeate channel",
-            default=tmp_dict,
-        )
-        self.mixed_permeate = self.config.property_package.state_block_class(
-            self.flowsheet().config.time,
-            doc="Material properties of mixed permeate exiting the module",
-            default=tmp_dict,
-        )
-        # Interface properties
-        self.feed_side.properties_interface = (
-            self.config.property_package.state_block_class(
-                self.flowsheet().config.time,
-                self.length_domain,
-                doc="Material properties of feed-side membrane interface",
-                default=tmp_dict,
-            )
-        )
-
-        # Add Ports
-        self.add_inlet_port(name="inlet", block=self.feed_side)
-        self.add_outlet_port(name="retentate", block=self.feed_side)
-        self.add_port(name="permeate", block=self.mixed_permeate)
-
-        # References for control volume
-        # pressure change
-        if (
-            self.config.has_pressure_change
-            and self.config.momentum_balance_type != "none"
-        ):
-            self.deltaP = Reference(self.feed_side.deltaP)
-
-        self._make_performance()
-
-        self._add_expressions()
-
-    def _make_performance(self):
-
-        units_meta = self.config.property_package.get_metadata().get_derived_units
-
-        solvent_set = self.config.property_package.solvent_set
-        solute_set = self.config.property_package.solute_set
-
-        if self.config.pressure_change_type == PressureChangeType.calculated:
-            pass
-        elif (
-            self.config.pressure_change_type == PressureChangeType.fixed_per_unit_length
-        ):
-            self.dP_dx = Var(
-                self.flowsheet().config.time,
-                initialize=-5e4,
-                bounds=(-2e5, -1e3),
-                domain=NegativeReals,
-                units=units_meta("pressure") * units_meta("length") ** -1,
-                doc="pressure drop per unit length across feed channel",
-            )
-
-        if (self.config.pressure_change_type != PressureChangeType.fixed_per_stage) or (
-            self.config.mass_transfer_coefficient == MassTransferCoefficient.calculated
-        ):
-            # comes from ControlVolume1D in 1DRO
-            self.length = Var(
-                initialize=10,
-                bounds=(0.1, 5e2),
-                domain=NonNegativeReals,
-                units=units_meta("length"),
-                doc="Effective membrane length",
-            )
-
-        super()._make_performance()
-
-        # mass transfer
-        def mass_transfer_phase_comp_initialize(b, t, p, j):
-            return value(
-                self.feed_side.properties_in[t].get_material_flow_terms("Liq", j)
-                * self.recovery_mass_phase_comp[t, "Liq", j]
-            )
-
-        self.mass_transfer_phase_comp = Var(
-            self.flowsheet().config.time,
-            self.config.property_package.phase_list,
-            self.config.property_package.component_list,
-            initialize=mass_transfer_phase_comp_initialize,
-            bounds=(0.0, 1e6),
-            domain=NonNegativeReals,
-            units=units_meta("mass") * units_meta("time") ** -1,
-            doc="Mass transfer to permeate",
-        )
-
-        # constraints for additional variables (i.e. variables not used in other constraints)
-
-        @self.Constraint(
-            self.flowsheet().config.time,
-            self.config.property_package.phase_list,
-            self.config.property_package.component_list,
-            doc="Mass transfer term",
-        )
-        def eq_mass_transfer_term(self, t, p, j):
-            return (
-                self.mass_transfer_phase_comp[t, p, j]
-                == -self.feed_side.mass_transfer_term[t, p, j]
-            )
-
-        # Different expression in 1DRO
-        @self.Constraint(
-            self.flowsheet().config.time,
-            self.config.property_package.phase_list,
-            self.config.property_package.component_list,
-            doc="Permeate production",
-        )
-        def eq_permeate_production(b, t, p, j):
-            return (
-                b.mixed_permeate[t].get_material_flow_terms(p, j)
-                == b.area * b.flux_mass_phase_comp_avg[t, p, j]
-            )
-
-        # Feed and permeate-side connection
-        @self.Constraint(
-            self.flowsheet().config.time,
-            self.config.property_package.phase_list,
-            self.config.property_package.component_list,
-            doc="Mass transfer from feed to permeate",
-        )
-        def eq_connect_mass_transfer(b, t, p, j):
-            return (
-                b.mixed_permeate[t].get_material_flow_terms(p, j)
-                == -b.feed_side.mass_transfer_term[t, p, j]
-            )
-
-
-        # # Permeate-side stateblocks
-        # Not in 1DRO
-        @self.Constraint(
-            self.flowsheet().config.time,
-            self.length_domain,
-            solute_set,
-            doc="Permeate mass fraction",
-        )
-        def eq_mass_frac_permeate(b, t, x, j):
-            return (
-                b.permeate_side[t, x].mass_frac_phase_comp["Liq", j]
-                * sum(
-                    self.flux_mass_phase_comp[t, x, "Liq", jj]
-                    for jj in self.config.property_package.component_list
-                )
-                == self.flux_mass_phase_comp[t, x, "Liq", j]
-            )
-
-        # not in 1DRO
-        @self.Constraint(
-            self.flowsheet().config.time, self.length_domain, doc="Permeate flowrate"
-        )
-        def eq_flow_vol_permeate(b, t, x):
-            return (
-                b.permeate_side[t, x].flow_vol_phase["Liq"]
-                == b.mixed_permeate[t].flow_vol_phase["Liq"]
-            )
-
-        if self.config.pressure_change_type == PressureChangeType.fixed_per_unit_length:
-            # Pressure change equation when dP/dx = user-specified constant,
-            @self.Constraint(
-                self.flowsheet().config.time, doc="pressure change due to friction"
-            )
-            def eq_pressure_change(b, t):
-                return b.deltaP[t] == b.dP_dx[t] * b.length
-
-        elif self.config.pressure_change_type == PressureChangeType.calculated:
-
-            # Average pressure change per unit length due to friction
-            @self.Expression(
-                self.flowsheet().config.time,
-                doc="expression for average pressure change per unit length due to friction",
-            )
-            def dP_dx_avg(b, t):
-                return 0.5 * sum(b.dP_dx[t, x] for x in b.length_domain)
-
-            # Pressure change equation
-            @self.Constraint(
-                self.flowsheet().config.time, doc="pressure change due to friction"
-            )
-            def eq_pressure_change(b, t):
-                return b.deltaP[t] == b.dP_dx_avg[t] * b.length
-
-    def _add_expressions(self):
-        super()._add_expressions()
-
-        @self.Expression(self.flowsheet().config.time, doc="Over pressure ratio")
-        def over_pressure_ratio(b, t):
-            return (
-                b.feed_side.properties_out[t].pressure_osm_phase["Liq"]
-                - b.permeate_side[t, 1.0].pressure_osm_phase["Liq"]
-            ) / b.feed_side.properties_out[t].pressure
 
     def initialize_build(
         blk,
